token

+++ controllers/blogs.js
@@ -1,11 +1,8 @@
-// const express = require('express')
 const blogsRouter = require('express').Router()
-const { response } = require('express')
-const { request } = require('../app')
-const { findById } = require('../models/blog')
 const Blog = require('../models/blog')
 const User = require('../models/user')
-// app.use(express.json())
+const jwt = require('jsonwebtoken')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1})
@@ -14,8 +11,12 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-
-  const user = await User.findById(body.userId)
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if(!token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     ...body,
@@ -29,15 +30,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if(!token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
-
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if(!token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = { ...body }
-
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
   response.json(updatedBlog.toJSON())
 })
